python_rag/requirements.txt
asyncpg==0.30.0
fastapi==0.115.12
loguru==0.7.3
pydantic==2.11.5
PyPDF2==3.0.1
pytest==8.3.5
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
tqdm==4.67.1
uvicorn==0.34.2
async-lru==2.0.5
python-docx==1.1.2
sentence-transformers==4.1.0
python-multipart==0.0.20
sentence-transformers==4.1.0
transformers==4.52.4
torch==2.7.0

# optional to generate sample data
datasets==3.6.0
